+++ android/app/src/main/java/io/coursepick/coursepick/presentation/routetfinder/RouteFinderChoiceDialogFragment.kt
@@ -7,7 +7,7 @@ import androidx.fragment.app.DialogFragment
 import io.coursepick.coursepick.databinding.DialogRouteFinderChoiceBinding
 import io.coursepick.coursepick.domain.Coordinate
 import io.coursepick.coursepick.presentation.getSerializableCompat
-import io.coursepick.coursepick.presentation.view.preference.CoursePickPreferences
+import io.coursepick.coursepick.presentation.preference.CoursePickPreferences
 
 class RouteFinderChoiceDialogFragment :
     DialogFragment(),
