+++ android/app/src/main/java/io/coursepick/coursepick/presentation/view/BindingAdapter.kt
@@ -1,4 +1,4 @@
-package io.coursepick.coursepick.presentation
+package io.coursepick.coursepick.presentation.view
 
 import android.content.Context
 import android.view.View
